bridge/requirements.txt
psutil>=5.9,<7
pyserial>=3.5,<4
